Add ConversationTranslationRecognizer class, make internal to ConversationTranslator, keep TranslationRecognizer mostly as it was before
